Create svm-boult-update.c
I compared libsvm's 3.1 and 3.20 with emacs ediff, while I had Dr. Boult's svm.c open on the side. For every difference found between 3.1 and 3.20, I compared it to Boult's file and made the necessary changes to his. For every addition I made, I commented above and below it with the word "UPDATE", so you could just do a keyword find to spot all the changes. Dr. Boult had a lot more functions in his program that were not present in the original svm.cpp, so I commented above those just in case changes needed to be made there as well. This was especially present in the svm_train category, as Dr. Boult had many more comparisons.